+++ demo/entities/src/main/java/software/xdev/sse/demo/entities/Product.java
@@ -3,16 +3,14 @@ package software.xdev.sse.demo.entities;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
+
+import software.xdev.sse.demo.service.validation.StringValidator;
 
 
 @Entity
 @Table(name = "product")
 public class Product extends IdentifiableEntity
 {
-	@NotNull
-	@Size(max = 255)
 	@Column(name = "name", nullable = false, unique = true)
 	private String name;
 	
@@ -22,7 +20,7 @@ public class Product extends IdentifiableEntity
 	
 	public Product(final String name)
 	{
-		this.name = name;
+		this.setName(name);
 	}
 	
 	public String getName()
@@ -32,6 +30,6 @@ public class Product extends IdentifiableEntity
 	
 	public void setName(final String name)
 	{
-		this.name = name;
+		this.name = StringValidator.notTooLongDefault("name", name);
 	}
 }